+++ mozilla_version/firefox.py
@@ -5,7 +5,7 @@ Examples:
 
         from mozilla_version.firefox import FirefoxVersion
 
-        version = FirefoxVersion('60.0.1')
+        version = FirefoxVersion.parse('60.0.1')
 
         version.major_number    # 60
         version.minor_number    # 0
@@ -17,7 +17,7 @@ Examples:
 
         str(version)        # '60.0.1'
 
-        previous_version = FirefoxVersion('60.0b14')
+        previous_version = FirefoxVersion.parse('60.0b14')
         previous_version < version      # True
 
         previous_version.beta_number    # 14
@@ -29,47 +29,101 @@ Examples:
         previous_version.is_release  # False
         previous_version.is_nightly  # False
 
-        invalid_version = FirefoxVersion('60.1')      # raises InvalidVersionError
-        invalid_version = FirefoxVersion('60.0.0')    # raises InvalidVersionError
-        version = FirefoxVersion('60.0')    # valid
+        invalid_version = FirefoxVersion.parse('60.1')      # raises InvalidVersionError
+        invalid_version = FirefoxVersion.parse('60.0.0')    # raises InvalidVersionError
+        version = FirefoxVersion.parse('60.0')    # valid
+
+        # Versions can be built by raw values
+        FirefoxVersion(60, 0))         # '60.0'
+        FirefoxVersion(60, 0, 1))      # '60.0.1'
+        FirefoxVersion(60, 1, 0))      # '60.1.0'
+        FirefoxVersion(60, 0, 1, 1))   # '60.0.1build1'
+        FirefoxVersion(60, 0, beta_number=1))       # '60.0b1'
+        FirefoxVersion(60, 0, is_nightly=True))     # '60.0a1'
+        FirefoxVersion(60, 0, is_aurora_or_devedition=True))    # '60.0a2'
+        FirefoxVersion(60, 0, is_esr=True))         # '60.0esr'
+        FirefoxVersion(60, 0, 1, is_esr=True))      # '60.0.1esr'
 
 """
 
 import re
+import attr
 
 from mozilla_version.errors import (
     InvalidVersionError, MissingFieldError, TooManyTypesError, NoVersionTypeError
 )
 from mozilla_version.version import VersionType
 
-_VALID_VERSION_PATTERN = re.compile(r"""
-^(?P<major_number>\d+)\.(
-    (?P<zero_minor_number>0)
-        (   # 2-digit-versions (like 46.0, 46.0b1, 46.0esr)
-            (?P<is_nightly>a1)
-            |(?P<is_aurora_or_devedition>a2)
-            |b(?P<beta_number>\d+)
-            |(?P<is_two_digit_esr>esr)
-        )?
-    |(  # Here begins the 3-digit-versions.
-        (?P<non_zero_minor_number>[1-9]\d*)\.(?P<potential_zero_patch_number>\d+)
-        |(?P<potential_zero_minor_number>\d+)\.(?P<non_zero_patch_number>[1-9]\d*)
-        # 46.0.0 is not correct
-    )(?P<is_three_digit_esr>esr)? # Neither is 46.2.0b1
-    # 3-digits end
-)(?P<has_build_number>build(?P<build_number>\d+))?$""", re.VERBOSE)
-# See more examples of (in)valid versions in the tests
+# XXX This pattern doesn't catch all subtleties of a Firefox version (like 32.5 isn't valid).
+# This regex is intended to assign numbers. Then checks are done by attrs and __attrs_post_init__()
+_VALID_ENOUGH_VERSION_PATTERN = re.compile(r"""
+^(?P<major_number>\d+)
+\.(?P<minor_number>\d+)
+(\.(?P<patch_number>\d+))?
+(
+    (?P<is_nightly>a1)
+    |(?P<is_aurora_or_devedition>a2)
+    |b(?P<beta_number>\d+)
+    |(?P<is_esr>esr)
+)?
+(build(?P<build_number>\d+))?$""", re.VERBOSE)
 
 
-_NUMBERS_TO_REGEX_GROUP_NAMES = {
-    'major_number': ('major_number',),
-    'minor_number': ('zero_minor_number', 'non_zero_minor_number', 'potential_zero_minor_number'),
-    'patch_number': ('non_zero_patch_number', 'potential_zero_patch_number'),
-    'beta_number': ('beta_number',),
-    'build_number': ('build_number',),
-}
+def _positive_int(val):
+    if isinstance(val, float):
+        raise ValueError('"{}" must not be a float'.format(val))
+    val = int(val)
+    if val >= 0:
+        return val
+    raise ValueError('"{}" must be positive'.format(val))
 
 
+def _positive_int_or_none(val):
+    if val is None:
+        return val
+    return _positive_int(val)
+
+
+def _strictly_positive_int_or_none(val):
+    val = _positive_int_or_none(val)
+    if val is None or val > 0:
+        return val
+    raise ValueError('"{}" must be strictly positive'.format(val))
+
+
+def _find_type(version):
+    version_type = None
+
+    def ensure_version_type_is_not_already_defined(previous_type, candidate_type):
+        if previous_type is not None:
+            raise TooManyTypesError(
+                str(version), previous_type, candidate_type
+            )
+
+    if version.is_nightly:
+        version_type = VersionType.NIGHTLY
+    if version.is_aurora_or_devedition:
+        ensure_version_type_is_not_already_defined(
+            version_type, VersionType.AURORA_OR_DEVEDITION
+        )
+        version_type = VersionType.AURORA_OR_DEVEDITION
+    if version.is_beta:
+        ensure_version_type_is_not_already_defined(version_type, VersionType.BETA)
+        version_type = VersionType.BETA
+    if version.is_esr:
+        ensure_version_type_is_not_already_defined(version_type, VersionType.ESR)
+        version_type = VersionType.ESR
+    if version.is_release:
+        ensure_version_type_is_not_already_defined(version_type, VersionType.RELEASE)
+        version_type = VersionType.RELEASE
+
+    if version_type is None:
+        raise NoVersionTypeError(str(version))
+
+    return version_type
+
+
+@attr.s(frozen=True, cmp=False)
 class FirefoxVersion(object):
     """Class that validates and handles Firefox version numbers.
 
@@ -80,107 +134,62 @@ class FirefoxVersion(object):
         InvalidVersionError: if the string doesn't match the pattern of a valid version number
         MissingFieldError: if a mandatory field is missing in the string. Mandatory fields are
             `major_number` and `minor_number`
-        TypeError: if an integer can't be cast from the string
+        ValueError: if an integer can't be cast or is not (strictly) positive
         TooManyTypesError: if the string matches more than 1 `VersionType`
         NoVersionTypeError: if the string matches none.
 
     """
 
-    def __init__(self, version_string):
+    major_number = attr.ib(type=int, converter=_positive_int)
+    minor_number = attr.ib(type=int, converter=_positive_int)
+    patch_number = attr.ib(type=int, converter=_positive_int_or_none, default=None)
+    build_number = attr.ib(type=int, converter=_strictly_positive_int_or_none, default=None)
+    beta_number = attr.ib(type=int, converter=_strictly_positive_int_or_none, default=None)
+    is_nightly = attr.ib(type=bool, default=False)
+    is_aurora_or_devedition = attr.ib(type=bool, default=False)
+    is_esr = attr.ib(type=bool, default=False)
+    version_type = attr.ib(init=False, default=attr.Factory(_find_type, takes_self=True))
+
+    def __attrs_post_init__(self):
+        """Ensure attributes are sane all together."""
+        if (
+            (self.minor_number == 0 and self.patch_number == 0) or
+            (self.minor_number != 0 and self.patch_number is None) or
+            (self.beta_number is not None and self.patch_number is not None) or
+            (self.patch_number is not None and self.is_nightly) or
+            (self.patch_number is not None and self.is_aurora_or_devedition)
+        ):
+            raise InvalidVersionError(self)
+
+    @classmethod
+    def parse(cls, version_string):
         """Construct an object representing a valid Firefox version number."""
-        self._version_string = version_string
-        self._regex_matches = _VALID_VERSION_PATTERN.match(self._version_string)
-        if self._regex_matches is None:
-            raise InvalidVersionError(self._version_string)
+        regex_matches = _VALID_ENOUGH_VERSION_PATTERN.match(version_string)
+        if regex_matches is None:
+            raise InvalidVersionError(version_string)
+
+        args = {}
 
         for field in ('major_number', 'minor_number'):
-            self._assign_mandatory_number(field)
+            args[field] = _get_value_matched_by_regex(field, regex_matches, version_string)
 
         for field in ('patch_number', 'beta_number', 'build_number'):
-            self._assign_optional_number(field)
+            try:
+                args[field] = _get_value_matched_by_regex(field, regex_matches, version_string)
+            except MissingFieldError:
+                pass
 
-        self._perform_sanity_checks()
-
-    def _assign_mandatory_number(self, field_name):
-        matched_value = _get_value_matched_by_regex(
-            field_name, self._regex_matches, self._version_string
+        return cls(
+            is_nightly=regex_matches.group('is_nightly') is not None,
+            is_aurora_or_devedition=regex_matches.group('is_aurora_or_devedition') is not None,
+            is_esr=regex_matches.group('is_esr') is not None,
+            **args
         )
-        setattr(self, field_name, int(matched_value))
-
-    def _assign_optional_number(self, field_name):
-        try:
-            self._assign_mandatory_number(field_name)
-        except (MissingFieldError, TypeError):    # TypeError is when None can't be cast to int
-            pass
-
-    def _perform_sanity_checks(self):
-        self._process_and_ensure_type_is_unique()
-
-    def _process_and_ensure_type_is_unique(self):
-        version_type = None
-
-        def ensure_version_type_is_not_already_defined(previous_type, candidate_type):
-            if previous_type is not None:
-                raise TooManyTypesError(
-                    self._version_string, previous_type, candidate_type
-                )
-
-        if self.is_nightly:
-            version_type = VersionType.NIGHTLY
-        if self.is_aurora_or_devedition:
-            ensure_version_type_is_not_already_defined(
-                version_type, VersionType.AURORA_OR_DEVEDITION
-            )
-            version_type = VersionType.AURORA_OR_DEVEDITION
-        if self.is_beta:
-            ensure_version_type_is_not_already_defined(version_type, VersionType.BETA)
-            version_type = VersionType.BETA
-        if self.is_esr:
-            ensure_version_type_is_not_already_defined(version_type, VersionType.ESR)
-            version_type = VersionType.ESR
-        if self.is_release:
-            ensure_version_type_is_not_already_defined(version_type, VersionType.RELEASE)
-            version_type = VersionType.RELEASE
-
-        if version_type is None:
-            raise NoVersionTypeError(self._version_string)
-
-        self.version_type = version_type
-
-    @property
-    def is_nightly(self):
-        """Return `True` if `FirefoxVersion` was built with a string matching a nightly version."""
-        return self._regex_matches.group('is_nightly') is not None
-
-    @property
-    def is_aurora_or_devedition(self):
-        """Return `True` if `FirefoxVersion` was built with a string matching an aurora version.
-
-        Aurora was later called Firefox DevEdition. Starting Firefox 55, DevEdition is shipped
-        with a beta version number. This means there is no version "X.0a2" with X >= 55.
-
-        """
-        # TODO raise error for major_number > X. X being the first release shipped after we moved
-        # devedition onto beta.
-        return self._regex_matches.group('is_aurora_or_devedition') is not None
 
     @property
     def is_beta(self):
         """Return `True` if `FirefoxVersion` was built with a string matching a beta version."""
-        try:
-            self.beta_number
-            return True
-        except AttributeError:
-            return False
-
-    @property
-    def is_esr(self):
-        """Return `True` if `FirefoxVersion` was built with a string matching an ESR version.
-
-        Firefox ESR stands for "Extended Support Release".
-        """
-        return self._regex_matches.group('is_two_digit_esr') is not None or \
-            self._regex_matches.group('is_three_digit_esr') is not None
+        return self.beta_number is not None
 
     @property
     def is_release(self):
@@ -190,9 +199,27 @@ class FirefoxVersion(object):
     def __str__(self):
         """Implement string representation.
 
-        Return the original string passed to the constructor.
+        Computes a new string based on the given attributes.
         """
-        return self._version_string
+        semvers = [str(self.major_number), str(self.minor_number)]
+        if self.patch_number is not None:
+            semvers.append(str(self.patch_number))
+
+        string = '.'.join(semvers)
+
+        if self.is_nightly:
+            string = '{}a1'.format(string)
+        elif self.is_aurora_or_devedition:
+            string = '{}a2'.format(string)
+        elif self.is_beta:
+            string = '{}b{}'.format(string, self.beta_number)
+        elif self.is_esr:
+            string = '{}esr'.format(string)
+
+        if self.build_number is not None:
+            string = '{}build{}'.format(string, self.build_number)
+
+        return string
 
     def __eq__(self, other):
         """Implement `==` operator.
@@ -205,18 +232,18 @@ class FirefoxVersion(object):
         Examples:
             .. code-block:: python
 
-                assert FirefoxVersion('60.0') == FirefoxVersion('60.0')
-                assert FirefoxVersion('60.0') == FirefoxVersion('60.0esr')
-                assert FirefoxVersion('60.0') == FirefoxVersion('60.0build1')
-                assert FirefoxVersion('60.0build1') == FirefoxVersion('60.0build1')
+                assert FirefoxVersion.parse('60.0') == FirefoxVersion.parse('60.0')
+                assert FirefoxVersion.parse('60.0') == FirefoxVersion.parse('60.0esr')
+                assert FirefoxVersion.parse('60.0') == FirefoxVersion.parse('60.0build1')
+                assert FirefoxVersion.parse('60.0build1') == FirefoxVersion.parse('60.0build1')
 
-                assert FirefoxVersion('60.0') != FirefoxVersion('61.0')
-                assert FirefoxVersion('60.0') != FirefoxVersion('60.1.0')
-                assert FirefoxVersion('60.0') != FirefoxVersion('60.0.1')
-                assert FirefoxVersion('60.0') != FirefoxVersion('60.0a1')
-                assert FirefoxVersion('60.0') != FirefoxVersion('60.0a2')
-                assert FirefoxVersion('60.0') != FirefoxVersion('60.0b1')
-                assert FirefoxVersion('60.0build1') != FirefoxVersion('60.0build2')
+                assert FirefoxVersion.parse('60.0') != FirefoxVersion.parse('61.0')
+                assert FirefoxVersion.parse('60.0') != FirefoxVersion.parse('60.1.0')
+                assert FirefoxVersion.parse('60.0') != FirefoxVersion.parse('60.0.1')
+                assert FirefoxVersion.parse('60.0') != FirefoxVersion.parse('60.0a1')
+                assert FirefoxVersion.parse('60.0') != FirefoxVersion.parse('60.0a2')
+                assert FirefoxVersion.parse('60.0') != FirefoxVersion.parse('60.0b1')
+                assert FirefoxVersion.parse('60.0build1') != FirefoxVersion.parse('60.0build2')
 
         """
         return self._compare(other) == 0
@@ -251,8 +278,10 @@ class FirefoxVersion(object):
 
         """
         for field in ('major_number', 'minor_number', 'patch_number'):
-            this_number = getattr(self, field, 0)
-            other_number = getattr(other, field, 0)
+            this_number = getattr(self, field)
+            this_number = 0 if this_number is None else this_number
+            other_number = getattr(other, field)
+            other_number = 0 if other_number is None else other_number
 
             difference = this_number - other_number
 
@@ -273,7 +302,7 @@ class FirefoxVersion(object):
         # we only compare build_numbers when we both have them.
         try:
             return self.build_number - other.build_number
-        except AttributeError:
+        except TypeError:
             pass
 
         return 0
@@ -283,13 +312,11 @@ class FirefoxVersion(object):
 
 
 def _get_value_matched_by_regex(field_name, regex_matches, version_string):
-    group_names = _NUMBERS_TO_REGEX_GROUP_NAMES[field_name]
-    for group_name in group_names:
-        try:
-            value = regex_matches.group(group_name)
-            if value is not None:
-                return value
-        except IndexError:
-            pass
+    try:
+        value = regex_matches.group(field_name)
+        if value is not None:
+            return value
+    except IndexError:
+        pass
 
     raise MissingFieldError(version_string, field_name)